Move Fill section to component
