+++ src/components/App.js
@@ -10,6 +10,7 @@ import callToApi from '../services/api';
 import Header from './Header';
 import CardPreview from './CardPreview';
 import Design from './Design';
+import Fill from './Fill';
 
 function App() {
   // /****VARIABLES****/
@@ -106,18 +107,6 @@ function App() {
 
   // /*****FUNCIONALIDADES*****/
 
-  const errorPhoneText = (errorMsg) => {
-    if (errorPhone) {
-      return errorMsg;
-    }
-  };
-
-  const errorEmailText = (errorMsg) => {
-    if (errorEmail) {
-      return errorMsg;
-    }
-  };
-
   // /*****END FUNCIONALIDADES*****/
 
   return (
@@ -134,150 +123,19 @@ function App() {
           </div>
         </section>
         <form className="container-form" onSubmit={handleSubmit}>
-        <Design handleInput={handleInput} handleClickDisign={handleClickDesign}palette={data.palette}  activeSection={activeSection}></Design>
-          <fieldset className="fill">
-            <div className="fill__title" onClick={handleClickFill}>
-              <legend className="fill__legend">
-                <i className="fa-regular fa-keyboard fill__legend--icon"></i>{' '}
-                Rellena
-              </legend>
-              <i className="fa-solid fa-angle-up collapse legend--arrow--up"></i>
-              <i className="fa-solid fa-angle-down legend--arrow--down"></i>
-            </div>
-
-            <div
-              className={`js-fill ${
-                activeSection !== 'fill' ? 'collapse' : ''
-              }`}>
-              <label htmlFor="name" className="fill__label fill__label--name">
-                <span className="fill__label__text--name">
-                  {' '}
-                  Nombre completo<span className="asterisc">*</span>
-                </span>
-                <input
-                  className="fill__input fill__input--name js-input js_input_name"
-                  type="text"
-                  name="name"
-                  id="name"
-                  placeholder="Ej: Sally Jill"
-                  onChange={handleInput}
-                  required
-                  value={data.name}
-                />
-              </label>
-              <label
-                htmlFor="job-position"
-                className="fill__label fill__label--jobposition">
-                <span className="fill__label__text--jobposition">
-                  Puesto<span className="asterisc">*</span>
-                </span>
-                <input
-                  className="fill__input fill__input--jobposition js-input js_input_job"
-                  type="text"
-                  name="job"
-                  id="job-position"
-                  placeholder="Ej: Front-end unicorn"
-                  onChange={handleInput}
-                  required
-                  value={data.job}
-                />
-              </label>
-              <label
-                htmlFor="profile-image"
-                className="fill__label fill__label--profileimage">
-                <div className="fill__label__container ">
-                  <label
-                    className="fill__label__container--button"
-                    htmlFor="img-selector">
-                    Añadir imagen
-                  </label>
-                  <input
-                    type="file"
-                    name="photo"
-                    id="img-selector"
-                    className="action__hiddenField js__profile-upload-btn js_input_photo fill__label__text--profileimage collapse"
-                    required
-                  />
-
-                  <div className="fill__label__container--box js__profile-preview"></div>
-                </div>
-              </label>
-              <label htmlFor="email" className="fill__label fill__label--email">
-                <span className="fill__label__text--email">
-                  Email<span className="asterisc">*</span>
-                </span>
-                <input
-                  className="fill__input fill__input--email js-input js_input_email"
-                  type="email"
-                  name="email"
-                  id="email"
-                  placeholder="Ej: sally.hill@gmail.com"
-                  onChange={handleInput}
-                  required
-                  value={data.email}
-                />
-              </label>
-              <p className="error-msg">
-                {' '}
-                {errorEmailText('El email que has introducido no es correcto.')}
-              </p>
-              <label
-                htmlFor="phone"
-                className="fill__label fill__input--telefono">
-                <span className="fill__label__text--telefono">Teléfono</span>
-                <input
-                  className="fill__input fill__input--telefono js-input js_input_phone"
-                  type="text"
-                  name="phone"
-                  id="phone"
-                  placeholder="Ej: 555-55-55-55"
-                  onChange={handleInput}
-                  value={data.phone}
-                />
-              </label>
-              <p className="error-msg">
-                {' '}
-                {errorPhoneText(
-                  'El teléfono que has introducido no es correcto.'
-                )}
-              </p>
-              <label
-                htmlFor="linkedin"
-                className="fill__label fill__label--linkedin">
-                <span className="fill__label__text--linkedin">
-                  Linkedin<span className="asterisc">*</span>
-                </span>
-                <input
-                  className="fill__input fill__input--linkedin js-input js_input_linkedin"
-                  type="text"
-                  name="linkedin"
-                  id="linkedin"
-                  placeholder="Ej: linkedin.com/in/sally-hill"
-                  onChange={handleInput}
-                  required
-                  value={data.linkedin}
-                />
-              </label>
-              <label
-                htmlFor="github"
-                className="fill__label fill__label--github">
-                <span className="fill__label__text--github">
-                  Github<span className="asterisc">*</span>
-                </span>
-                <input
-                  className="fill__input fill__input--github js-input js_input_github"
-                  type="text"
-                  name="github"
-                  id="github"
-                  placeholder="Ej: @sally-hill"
-                  onChange={handleInput}
-                  required
-                  value={data.github}
-                />
-              </label>
-            </div>
-            <div className="border_button"></div>
-          </fieldset>
+          <Design
+            handleInput={handleInput}
+            handleClickDesign={handleClickDesign}
+            palette={data.palette}
+            activeSection={activeSection}></Design>
+          <Fill
+            handleInput={handleInput}
+            handleClickFill={handleClickFill}
+            data={data}
+            activeSection={activeSection}
+            errorPhone={errorPhone}
+            errorEmail={errorEmail}
+          />
           <fieldset className="form">
             <div className="form__share" onClick={handleClickShare}>
               <legend className="form__legend">
